+++ lib/src/core/extended_context.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/widgets.dart';
-import 'package:vrouter/src/core/vrouter_data.dart';
 import 'package:vrouter/src/widgets/vrouter.dart';
+
+import 'vrouter_sailor/vrouter_sailor.dart';
 
 extension VRouterContext on BuildContext {
   InitializedVRouterSailor get vRouter => VRouter.of(this);