+++ lib/src/vroute_elements/vnester_base.dart
@@ -101,8 +101,9 @@ class VNesterBase extends VRouteElementBuilder {
   ///
   /// Also see:
   ///   * [VRouter.buildTransition] for default transitions for all routes
-  final Widget Function(Animation<double> animation,
-      Animation<double> secondaryAnimation, Widget child)? buildTransition;
+  final Widget Function(
+          Animation<double> animation, Animation<double> secondaryAnimation, Widget child)?
+      buildTransition;
 
   /// A key for the nested navigator
   /// It is created automatically
@@ -134,6 +135,36 @@ class VNesterBase extends VRouteElementBuilder {
     this.fullscreenDialog = false,
   });
 
+  /// Provides a [state] from which to access [VRouter] data in [widgetBuilder]
+  VNesterBase.builder({
+    required Widget Function(BuildContext context, VRouterData state, Widget child)
+        widgetBuilder,
+    required List<VRouteElement> nestedRoutes,
+    Duration? transitionDuration,
+    Duration? reverseTransitionDuration,
+    Widget Function(
+            Animation<double> animation, Animation<double> secondaryAnimation, Widget child)?
+        buildTransition,
+    LocalKey? key,
+    String? name,
+    List<VRouteElement> stackedRoutes = const [],
+    GlobalKey<NavigatorState>? navigatorKey,
+    fullscreenDialog = false,
+  }) : this(
+          widgetBuilder: (child) => VRouterDataBuilder(
+            builder: (context, state) => widgetBuilder(context, state, child),
+          ),
+          nestedRoutes: nestedRoutes,
+          transitionDuration: transitionDuration,
+          reverseTransitionDuration: reverseTransitionDuration,
+          buildTransition: buildTransition,
+          key: key,
+          name: name,
+          stackedRoutes: stackedRoutes,
+          navigatorKey: navigatorKey,
+          fullscreenDialog: fullscreenDialog,
+        );
+
   @override
   List<VRouteElement> buildRoutes() => [
         VNesterPageBase(
@@ -143,8 +174,7 @@ class VNesterBase extends VRouteElementBuilder {
           stackedRoutes: stackedRoutes,
           widgetBuilder: widgetBuilder,
           navigatorKey: navigatorKey,
-          pageBuilder: (LocalKey key, Widget child, String? name) =>
-              VDefaultPage.fromPlatform(
+          pageBuilder: (LocalKey key, Widget child, String? name, VRouterData state) => VDefaultPage.fromPlatform(
             key: key,
             child: child,
             name: name,