+++ lib/src/vroute_elements/vnester_page_base.dart
@@ -45,7 +45,7 @@ class VNesterPageBase extends VRouteElement
   /// Function which returns a page that will wrap [widget]
   ///   - key and name should be given to your [Page]
   ///   - child should be placed as the last child in [Route]
-  final Page Function(LocalKey key, Widget child, String? name) pageBuilder;
+  final Page Function(LocalKey key, Widget child, String? name, VRouterData state) pageBuilder;
 
   /// A key for the nested navigator
   /// It is created automatically
@@ -72,6 +72,28 @@ class VNesterPageBase extends VRouteElement
         );
   }
 
+  /// Provides a [state] from which to access [VRouter] data in [widgetBuilder]
+  VNesterPageBase.builder({
+    required List<VRouteElement> nestedRoutes,
+    required Widget Function(BuildContext context, VRouterData state, Widget child)
+        widgetBuilder,
+    required Page Function(LocalKey key, Widget child, String? name) pageBuilder,
+    List<VRouteElement> stackedRoutes = const [],
+    LocalKey? key,
+    String? name,
+    GlobalKey<NavigatorState>? navigatorKey,
+  }) : this(
+          nestedRoutes: nestedRoutes,
+          widgetBuilder: (child) => VRouterDataBuilder(
+            builder: (context, state) => widgetBuilder(context, state, child),
+          ),
+          pageBuilder: (key, child, name, _) => pageBuilder(key, child, name),
+          stackedRoutes: stackedRoutes,
+          key: key,
+          name: name,
+          navigatorKey: navigatorKey,
+        );
+
   /// A hero controller for the navigator
   /// It is created automatically
   final HeroController heroController = HeroController();
@@ -94,8 +116,12 @@ class VNesterPageBase extends VRouteElement
             remainingPath: parentVPathMatch.remainingPath,
             pathParameters: parentVPathMatch.pathParameters,
             localPath: null,
+            names: parentVPathMatch.names + [if (name != null) name!],
           )
-        : InvalidVPathMatch(localPath: null);
+        : InvalidVPathMatch(
+            localPath: null,
+            names: parentVPathMatch.names + [if (name != null) name!],
+          );
 
     // Try to find valid VRoute from nestedRoutes
     VRoute? nestedRouteVRoute;
@@ -139,6 +165,11 @@ class VNesterPageBase extends VRouteElement
       ...nestedRouteVRoute.pathParameters,
       ...stackedRouteVRoute?.pathParameters ?? {},
     };
+
+    final names = <String>[
+      ...nestedRouteVRoute.names,
+      ...stackedRouteVRoute?.names ?? [],
+    ].toSet().toList();
 
     return VRoute(
       vRouteElementNode: vRouteElementNode,
@@ -198,8 +229,7 @@ class VNesterPageBase extends VRouteElement
                               vPopData = data;
                             } else {
                               vPopData = VPopData(
-                                elementToPop: nestedRouteVRoute!
-                                    .vRouteElementNode
+                                elementToPop: nestedRouteVRoute!.vRouteElementNode
                                     .getVRouteElementToPop(),
                                 pathParameters: pathParameters,
                                 queryParameters: {},
@@ -240,6 +270,14 @@ class VNesterPageBase extends VRouteElement
             ),
           ),
           name ?? parentVPathMatch.localPath,
+          VRouterDataImpl(
+            previousUrl: vPathRequestData.previousUrl,
+            url: vPathRequestData.url,
+            pathParameters: pathParameters,
+            queryParameters: vPathRequestData.queryParameters,
+            historyState: vPathRequestData.historyState,
+            names: names,
+          ),
         ),
         ...stackedRouteVRoute?.pages ?? [],
       ],
@@ -247,6 +285,7 @@ class VNesterPageBase extends VRouteElement
       vRouteElements: <VRouteElement>[this] +
           nestedRouteVRoute.vRouteElements +
           (stackedRouteVRoute?.vRouteElements ?? []),
+      names: names,
     );
   }
 
@@ -304,8 +343,7 @@ class VNesterPageBase extends VRouteElement
                 values: [
                   OverlyPathParamsError(
                     pathParams: pathParameters.keys.toList(),
-                    expectedPathParams:
-                        parentPathResult.pathParameters.keys.toList(),
+                    expectedPathParams: parentPathResult.pathParameters.keys.toList(),
                   ),
                 ],
               ),
@@ -324,8 +362,7 @@ class VNesterPageBase extends VRouteElement
               MissingPathParamsError(
                 pathParams: pathParameters.keys.toList(),
                 missingPathParams:
-                    (parentPathResult as PathParamsErrorNewParentPath)
-                        .pathParameters,
+                    (parentPathResult as PathParamsErrorNewParentPath).pathParameters,
               ),
             ],
           ),
@@ -354,8 +391,8 @@ class VNesterPageBase extends VRouteElement
     }
 
     // Else try to find a NullPathError
-    if (childNameResults.indexWhere(
-            (childNameResult) => childNameResult is NullPathErrorNameResult) !=
+    if (childNameResults
+            .indexWhere((childNameResult) => childNameResult is NullPathErrorNameResult) !=
         -1) {
       return NullPathErrorNameResult(name: nameToMatch);
     }
@@ -414,8 +451,7 @@ class VNesterPageBase extends VRouteElement
             return ValidPopResult(
               path: childPopResult.path,
               poppedVRouteElements: childPopResult.poppedVRouteElements,
-              names:
-                  (name != null ? [name!] : <String>[]) + childPopResult.names,
+              names: (name != null ? [name!] : <String>[]) + childPopResult.names,
             );
           }
 
@@ -424,8 +460,7 @@ class VNesterPageBase extends VRouteElement
 
         // We should NOT pop with the VRouteElement to pop
 
-        final poppedVRouteElements =
-            childPopResult.poppedVRouteElements + [this];
+        final poppedVRouteElements = childPopResult.poppedVRouteElements + [this];
 
         // Check whether the parentPathResult is valid or not
         if (parentPathResult is ValidParentPathResult) {
@@ -446,8 +481,7 @@ class VNesterPageBase extends VRouteElement
             MissingPathParamsError(
               pathParams: pathParameters.keys.toList(),
               missingPathParams:
-                  (parentPathResult as PathParamsErrorNewParentPath)
-                      .pathParameters,
+                  (parentPathResult as PathParamsErrorNewParentPath).pathParameters,
             ),
           ],
         );
@@ -470,8 +504,7 @@ class VNesterPageBase extends VRouteElement
 class VNestedObserverReporter extends NavigatorObserver {
   final List<NavigatorObserver> _navigatorObserversToReportTo;
 
-  VNestedObserverReporter(
-      {required List<NavigatorObserver> navigatorObserversToReportTo})
+  VNestedObserverReporter({required List<NavigatorObserver> navigatorObserversToReportTo})
       : _navigatorObserversToReportTo = navigatorObserversToReportTo;
 
   @override